fix(Build): Fix build pipelines. npm run start works now
