+++ package.json
@@ -26,7 +26,7 @@
     "typescript": "^3.7.5"
   },
   "scripts": {
-    "start": "node src/server.js",
+    "start": "node server.js",
     "dev": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
