+++ server.js
@@ -1,0 +1,22 @@
+const path = require("path");
+const express = require("express");
+const app = express();
+const port = process.env.PORT || 3001;
+const publicPath = path.join(__dirname, "..", "public");
+
+app.use(express.static(publicPath));
+app.use(express.static(__dirname));
+app.use(express.static(path.join(__dirname, "build")));
+
+app.get("/ping", function(req, res) {
+  return res.send("pong");
+});
+app.get("/*", function(req, res) {
+  res.sendFile(path.join(__dirname, "build", "index.html"));
+});
+
+app.listen(port, () => {
+  console.log(__dirname);
+  console.log(path.join(publicPath, "index.html"));
+  console.log(`Server is up on port ${port}!`);
+});
